backend/requirements.txt
fastapi==0.115.0
uvicorn==0.31.0
anthropic==0.39.0
web3==7.4.0
python-dotenv==1.0.1
httpx==0.27.2
pydantic==2.9.2
pydantic-settings==2.5.2
apscheduler==3.10.4
